add cpu/memory
