+++ go.mod
@@ -8,7 +8,6 @@ require (
 	github.com/k0kubun/pp v3.0.1+incompatible
 	github.com/mattn/go-colorable v0.1.7 // indirect
 	github.com/mattn/go-isatty v0.0.12
-	github.com/pbnjay/memory v0.0.0-20201129165224-b12e5d931931
 	github.com/pkg/errors v0.9.1
 	github.com/shirou/gopsutil v3.20.11-0.20201116082039-2fb5da2f2449+incompatible
 	github.com/stretchr/testify v1.6.1
