+++ runtime/sys_test.go
@@ -1,22 +1,32 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 package gxruntime
 
 import (
-    "os"
 	"testing"
 	"time"
 )
 
-// exists returns whether the given file or directory exists
-func exists(path string) (bool, error) {
-	_, err := os.Stat(path)
-	if err == nil { return true, nil }
-	if os.IsNotExist(err) { return false, nil }
-	return false, err
-}
 
-func t1(t *testing.T) {
-	t.Logf("current os cpu number %d, memory limit %d bytes", GetCPUNum(), GetMemoryLimit())
+func TestSysStat(t *testing.T) {
+	t.Logf("current os cpu number %d", GetCPUNum())
+	total, used, free, usedPercent := GetMemoryStat()
+	t.Logf("memory: limit %d bytes, used %d bytes, free %d bytes, usedPercent %f", total, used, free, usedPercent)
 	t.Logf("current prcess thread number %d", GetThreadNum())
 	go func() {
 		time.Sleep(10e9)
@@ -53,7 +63,7 @@ func t1(t *testing.T) {
 	}
 	t.Logf("process memory usage percent %v", memoryUsage)
 
-	if ok, _ := exists(cgroupMemLimitPath); ok {
+	if IsCgroup() {
 		memoryLimit, err := GetCgroupMemoryLimit()
 		if err != nil {
 			t.Errorf("GetCgroupMemoryLimit() = error %+v", err)
@@ -65,7 +75,6 @@ func t1(t *testing.T) {
 			t.Errorf("GetCgroupProcessMemoryPercent(ps:%d) = error %+v", CurrentPID, err)
 		}
 		t.Logf("GetCgroupProcessMemoryPercent(ps:%d) = %+v", CurrentPID, memoryPercent)
-
 	}
 }
